python/x402/pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0,<1.22.0"]
build-backend = "hatchling.build"

[project]
name = "x402-tron"
version = "0.2.1"
description = "x402 Payment Protocol SDK for Python"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "x402", email = "dev@x402.io" }
]
keywords = ["x402", "payment", "blockchain", "tron"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "base58>=2.1.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.uv.sources]
agent-wallet = { git = "https://github.com/bankofai/agent-wallet.git", branch = "dev", subdirectory = "python" }

[project.optional-dependencies]
tron = ["tronpy>=0.4.0"]
fastapi = ["fastapi>=0.100.0"]
flask = ["flask>=2.0.0"]
agent-wallet = ["agent-wallet"]
all = [
    "tronpy>=0.4.0",
    "fastapi>=0.100.0",
    "flask>=2.0.0",
    "agent-wallet",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/bankofai/x402-tron"
Documentation = "https://github.com/bankofai/x402-tron#readme"
Repository = "https://github.com/bankofai/x402-tron"

[tool.hatch.build.targets.wheel]
packages = ["src/x402_tron"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
